refactor: move <ErrorBoundary /> to components dir
